fixes to corporate voidsuits & vaurca sprites

+++ code/game/machinery/suit_cycler.dm
@@ -45,6 +45,8 @@
 	var/can_repair			// If set, the cycler can repair voidsuits.
 	var/electrified = FALSE
 
+	//Can this cycler change departments?
+	var/department_change = TRUE
 	//Departments that the cycler can paint suits to look like.
 	var/list/departments = list("Engineering", "Mining", "Medical", "Security", "Atmos")
 	//Species that the suits can be configured to fit.
@@ -75,10 +77,11 @@
 			mask = new mask(src)
 
 	wires = new(src)
-	target_department = departments[1]
+	if(department_change)
+		target_department = departments[1]
 	target_species = species[1]
 	update_icon()
-	if(!target_department || !target_species)
+	if((!target_department && department_change) || !target_species)
 		qdel(src)
 
 /obj/machinery/suit_cycler/Destroy()
@@ -258,6 +261,7 @@
 	//Clear the access reqs, disable the safeties, and open up all paintjobs.
 	to_chat(user, SPAN_WARNING("You run the sequencer across the interface, corrupting the operating protocols."))
 	departments = list("Engineering", "Mining", "Medical", "Security", "Atmos", "^%###^%$")
+	department_change = TRUE
 	emagged = TRUE
 	safeties = FALSE
 	req_access = list()
@@ -296,6 +300,7 @@
 		"model_text" = model_text,
 		"radiation_level" = radiation_level,
 		"target_department" = target_department,
+		"department_change" = department_change,
 		"target_species" = target_species,
 		"helmet" = helmet ? list("name" = helmet.name, "damage" = 0) : null,
 		"suit" = suit ? list("name" = suit.name, "damage" = suit.damage) : null,
@@ -345,6 +350,8 @@
 		update_icon()
 
 	else if(action == "select_department")
+		if(!department_change)
+			return
 		var/choice = tgui_input_list(usr, "Please select the target department paintjob.", "Cycler Target Department Selection", departments, target_department)
 		if(choice)
 			target_department = choice
@@ -522,7 +529,7 @@
 
 //There HAS to be a less bloated way to do this. TODO: some kind of table/icon name coding? ~Z
 /obj/machinery/suit_cycler/proc/apply_paintjob()
-	if(!target_species || !target_department)
+	if(!target_species || (!target_department && department_change))
 		return
 
 	if(target_species)
@@ -531,102 +538,115 @@
 		if(suit && !suit.contained_sprite)
 			suit.refit_for_species(target_species)
 
-	switch(target_department)
-		if("Engineering")
-			if(helmet)
-				helmet.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
-				helmet.name = "engineering voidsuit helmet"
-				helmet.icon_state = "engineering_helm"
-				helmet.item_state = "engineering_helm"
-			if(suit)
-				suit.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
-				suit.name = "engineering voidsuit"
-				suit.icon_state = "engineering"
-				suit.item_state = "engineering"
-		if("Mining")
-			if(helmet)
-				helmet.name = "mining voidsuit helmet"
-				helmet.icon_state = "rig0-mining"
-				helmet.item_state = "mining_helm"
-			if(suit)
-				suit.name = "mining voidsuit"
-				suit.icon_state = "rig-mining"
-				suit.item_state = "mining_voidsuit"
-		if("Medical")
-			if(helmet)
-				helmet.name = "medical voidsuit helmet"
-				helmet.icon_state = "rig0-medical"
-				helmet.item_state = "medical_helm"
-			if(suit)
-				suit.name = "medical voidsuit"
-				suit.icon_state = "rig-medical"
-				suit.item_state = "medical_voidsuit"
-		if("Security")
-			if(helmet)
-				helmet.name = "security voidsuit helmet"
-				helmet.icon_state = "rig0-sec"
-				helmet.item_state = "sec_helm"
-			if(suit)
-				suit.name = "security voidsuit"
-				suit.icon_state = "rig-sec"
-				suit.item_state = "sec_voidsuit"
-		if("Atmos")
-			if(helmet)
-				helmet.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
-				helmet.name = "atmospherics voidsuit helmet"
-				helmet.icon_state = "atmos_helm"
-				helmet.item_state = "atmos_helm"
-			if(suit)
-				helmet.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
-				suit.name = "atmospherics voidsuit"
-				suit.icon_state = "atmos"
-				suit.item_state = "atmos"
-		if("Captain")
-			if(helmet)
-				helmet.name = "captain voidsuit helmet"
-				helmet.icon_state = "capspace"
-				helmet.item_state = "capspace"
-			if(suit)
-				suit.name = "captain voidsuit"
-				suit.icon_state = "capspace"
-				suit.item_state = "capspace"
-		if("^%###^%$", "Mercenary")
-			if(helmet)
-				helmet.name = "blood-red voidsuit helmet"
-				helmet.icon_state = "rig0-syndie"
-				helmet.item_state = "syndie_helm"
-			if(suit)
-				suit.name = "blood-red voidsuit"
-				suit.item_state = "syndie_voidsuit"
-				suit.icon_state = "rig-syndie"
-		if("Research")
-			if(helmet)
-				helmet.name = "research voidsuit helmet"
-				helmet.icon_state = "rig0-sci"
-				helmet.item_state = "research_voidsuit_helmet"
-			if(suit)
-				suit.name = "research voidsuit"
-				suit.item_state = "research_voidsuit"
-				suit.icon_state = "rig-sci"
+	if(department_change)
+		switch(target_department)
+			if("Engineering")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
+					helmet.name = "engineering voidsuit helmet"
+					helmet.icon_state = "engineering_helm"
+					helmet.item_state = "engineering_helm"
+				if(suit)
+					suit.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
+					suit.name = "engineering voidsuit"
+					suit.icon_state = "engineering"
+					suit.item_state = "engineering"
+			if("Mining")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/mining.dmi'
+					helmet.name = "mining voidsuit helmet"
+					helmet.icon_state = "mining_helm"
+					helmet.item_state = "mining_helm"
+				if(suit)
+					suit.icon = 'icons/obj/clothing/voidsuit/station/mining.dmi'
+					suit.name = "mining voidsuit"
+					suit.icon_state = "mining"
+					suit.item_state = "mining"
+			if("Medical")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/medical.dmi'
+					helmet.name = "medical voidsuit helmet"
+					helmet.icon_state = "medical_helm"
+					helmet.item_state = "medical_helm"
+				if(suit)
+					suit.icon = 'icons/obj/clothing/voidsuit/station/medical.dmi'
+					suit.name = "medical voidsuit"
+					suit.icon_state = "medical"
+					suit.item_state = "medical"
+			if("Security")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/security.dmi'
+					helmet.name = "security voidsuit helmet"
+					helmet.icon_state = "security_helm"
+					helmet.item_state = "security_helm"
+				if(suit)
+					suit.icon = 'icons/obj/clothing/voidsuit/station/security.dmi'
+					suit.name = "security voidsuit"
+					suit.icon_state = "security"
+					suit.item_state = "security"
+			if("Atmos")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
+					helmet.name = "atmospherics voidsuit helmet"
+					helmet.icon_state = "atmos_helm"
+					helmet.item_state = "atmos_helm"
+				if(suit)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/engineering.dmi'
+					suit.name = "atmospherics voidsuit"
+					suit.icon_state = "atmos"
+					suit.item_state = "atmos"
+			if("Captain")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/captain.dmi'
+					helmet.name = "captain voidsuit helmet"
+					helmet.icon_state = "captain_helm"
+					helmet.item_state = "captain_helm"
+				if(suit)
+					suit.icon = 'icons/obj/clothing/voidsuit/station/captain.dmi'
+					suit.name = "captain voidsuit"
+					suit.icon_state = "captain"
+					suit.item_state = "captain"
+			if("^%###^%$", "Mercenary")
+				if(helmet)
+					helmet.name = "blood-red voidsuit helmet"
+					helmet.icon_state = "rig0-syndie"
+					helmet.item_state = "syndie_helm"
+				if(suit)
+					suit.name = "blood-red voidsuit"
+					suit.item_state = "syndie_voidsuit"
+					suit.icon_state = "rig-syndie"
+			if("Research")
+				if(helmet)
+					helmet.icon = 'icons/obj/clothing/voidsuit/station/captain.dmi'
+					helmet.name = "research voidsuit helmet"
+					helmet.icon_state = "research_helm"
+					helmet.item_state = "research_helm"
+				if(suit)
+					suit.icon = 'icons/obj/clothing/voidsuit/station/captain.dmi'
+					suit.name = "research voidsuit"
+					suit.item_state = "research"
+					suit.icon_state = "research"
 
-		if("Freelancers")
-			if(helmet)
-				helmet.name = "freelancer voidsuit helmet"
-				helmet.icon_state = "rig0-freelancer"
-				helmet.item_state = "rig0-freelancer"
-			if(suit)
-				suit.name = "freelancer voidsuit"
-				suit.item_state = "freelancer"
-				suit.icon_state = "freelancer"
+			if("Freelancers")
+				if(helmet)
+					helmet.name = "freelancer voidsuit helmet"
+					helmet.icon_state = "rig0-freelancer"
+					helmet.item_state = "rig0-freelancer"
+				if(suit)
+					suit.name = "freelancer voidsuit"
+					suit.item_state = "freelancer"
+					suit.icon_state = "freelancer"
 
 	if(helmet)
 		if(helmet.contained_sprite)
 			helmet.refit_contained(target_species)
-		helmet.name = "refitted [helmet.name]"
+		if(department_change)
+			helmet.name = "refitted [helmet.name]"
 	if(suit)
 		if(suit.contained_sprite)
 			suit.refit_contained(target_species)
-		suit.name = "refitted [suit.name]"
+		if(department_change)
+			suit.name = "refitted [suit.name]"
 	update_icon()
 
 #undef TRY_INSERT_SUIT_PIECE
